Restructured storm package
