+++ src/main/scala/com/alawresz/twitterment/storm/akkatopology/InSpout.scala
@@ -1,4 +1,4 @@
-package com.alawresz.twitterment.storm.spouts
+package com.alawresz.twitterment.storm.akkatopology
 
 import com.alawresz.twitterment.configuration.ConsConfig
 import com.alawresz.twitterment.TweetModel.Tweet
